commit9

+++ customer-service/src/main/java/com/yigitusq/customer_service/dto/DtoCustomer.java
@@ -8,9 +8,15 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class DtoCustomer {
-    private int id;
+
+    private Long id;
+
     private String name;
+
     private String surname;
+
     private String email;
+
     private String status;
+
 }
